Fixing non-zero exit code in cachyos-reboot-required script (#13)
When reinstalling filesystem progs where is a check for mounted filesystems. If one does not have, for example, any XFS filesystem mounted when reinstalling, or upgrading xfsprogs the script produces an error in pacman execution. To avoid need to add "|| :" in if..then..else statement to do nothing instead of exiting the script with non-zero code.